requirements.txt
flask
flask-cors

# GitHub Copilot SDK Demo - Requirements
github-copilot-sdk
pydantic

# WhatsApp integration via Twilio
twilio

#teams
requests
